+++ src/Symfony/Bundle/FrameworkBundle/Command/SecretsListCommand.php
@@ -11,27 +11,33 @@
 
 namespace Symfony\Bundle\FrameworkBundle\Command;
 
-use Symfony\Bundle\FrameworkBundle\Exception\EncryptionKeyNotFoundException;
-use Symfony\Bundle\FrameworkBundle\Secret\Storage\SecretStorageInterface;
+use Symfony\Bundle\FrameworkBundle\Secrets\AbstractVault;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Helper\Dumper;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\ConsoleOutputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
 /**
  * @author Tobias Schultze <http://tobion.de>
  * @author Jérémy Derussé <jeremy@derusse.com>
+ * @author Nicolas Grekas <p@tchwork.com>
+ *
+ * @internal
  */
 final class SecretsListCommand extends Command
 {
-    protected static $defaultName = 'debug:secrets';
+    protected static $defaultName = 'secrets:list';
 
-    private $secretStorage;
+    private $vault;
+    private $localVault;
 
-    public function __construct(SecretStorageInterface $secretStorage)
+    public function __construct(AbstractVault $vault, AbstractVault $localVault = null)
     {
-        $this->secretStorage = $secretStorage;
+        $this->vault = $vault;
+        $this->localVault = $localVault;
 
         parent::__construct();
     }
@@ -39,54 +45,64 @@ final class SecretsListCommand extends Command
     protected function configure()
     {
         $this
-            ->setDefinition([
-                new InputOption('reveal', 'r', InputOption::VALUE_NONE, 'Display decrypted values alongside names'),
-            ])
             ->setDescription('Lists all secrets.')
+            ->addOption('reveal', 'r', InputOption::VALUE_NONE, 'Display decrypted values alongside names')
             ->setHelp(<<<'EOF'
 The <info>%command.name%</info> command list all stored secrets.
 
-    %command.full_name%
+    <info>%command.full_name%</info>
 
-When the the option <info>--reveal</info> is provided, the decrypted secrets are also displayed. 
+When the option <info>--reveal</info> is provided, the decrypted secrets are also displayed.
 
-    %command.full_name% --reveal
+    <info>%command.full_name% --reveal</info>
 EOF
             )
         ;
-
-        $this
-            ->setDescription('Lists all secrets.')
-            ->addOption('reveal', 'r', InputOption::VALUE_NONE, 'Display decrypted values alongside names');
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $reveal = $input->getOption('reveal');
-        $io = new SymfonyStyle($input, $output);
+        $io = new SymfonyStyle($input, $output instanceof ConsoleOutputInterface ? $output->getErrorOutput() : $output);
 
-        try {
-            $secrets = $this->secretStorage->listSecrets($reveal);
-        } catch (EncryptionKeyNotFoundException $e) {
-            throw new \LogicException(sprintf('Unable to decrypt secrets, the encryption key "%s" is missing.', $e->getKeyLocation()));
+        $io->comment('Use <info>"%env(secret:<name>)%"</info> to reference a secret in a config file.');
+
+        if (!$reveal = $input->getOption('reveal')) {
+            $io->comment(sprintf('To reveal the secrets run <info>php %s %s --reveal</info>', $_SERVER['PHP_SELF'], $this->getName()));
         }
 
-        if ($reveal) {
-            $rows = [];
-            foreach ($secrets as $name => $value) {
-                $rows[] = [$name, $value];
-            }
-            $io->table(['name', 'secret'], $rows);
-
-            return;
-        }
+        $secrets = $this->vault->list($reveal);
+        $localSecrets = null !== $this->localVault ? $this->localVault->list($reveal) : null;
 
         $rows = [];
-        foreach ($secrets as $name => $_) {
-            $rows[] = [$name];
+
+        $dump = new Dumper($output);
+        $dump = static function (?string $v) use ($dump) {
+            return null === $v ? '******' : $dump($v);
+        };
+
+        foreach ($secrets as $name => $value) {
+            $rows[$name] = [$name, $dump($value)];
         }
 
-        $io->comment(sprintf('To reveal the values of the secrets use <info>php %s %s --reveal</info>', $_SERVER['PHP_SELF'], $this->getName()));
-        $io->table(['name'], $rows);
+        if (null !== $message = $this->vault->getLastMessage()) {
+            $io->comment($message);
+        }
+
+        foreach ($localSecrets ?? [] as $name => $value) {
+            $rows[$name] = [$name, $rows[$name][1] ?? '', $dump($value)];
+        }
+
+        uksort($rows, 'strnatcmp');
+
+        if (null !== $this->localVault && null !== $message = $this->localVault->getLastMessage()) {
+            $io->comment($message);
+        }
+
+        (new SymfonyStyle($input, $output))
+            ->table(['Secret', 'Value'] + (null !== $localSecrets ? [2 => 'Local Value'] : []), $rows);
+
+        $io->comment("Local values override secret values.\nUse <info>secrets:set --local</info> to defined them.");
+
+        return 0;
     }
 }