+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/FrameworkExtension.php
@@ -25,7 +25,6 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Bundle\FrameworkBundle\Routing\AnnotatedRouteControllerLoader;
 use Symfony\Bundle\FrameworkBundle\Routing\RedirectableUrlMatcher;
 use Symfony\Bundle\FrameworkBundle\Routing\RouteLoaderInterface;
-use Symfony\Bundle\FrameworkBundle\Secret\Storage\SecretStorageInterface;
 use Symfony\Bundle\FullStack;
 use Symfony\Component\Asset\PackageInterface;
 use Symfony\Component\BrowserKit\AbstractBrowser;
@@ -1446,23 +1445,29 @@ class FrameworkExtension extends Extension
     private function registerSecretsConfiguration(array $config, ContainerBuilder $container, XmlFileLoader $loader)
     {
         if (!$this->isConfigEnabled($container, $config)) {
-            $container->removeDefinition('console.command.secrets_add');
+            $container->removeDefinition('console.command.secrets_set');
             $container->removeDefinition('console.command.secrets_list');
             $container->removeDefinition('console.command.secrets_remove');
             $container->removeDefinition('console.command.secrets_generate_key');
+            $container->removeDefinition('console.command.secrets_decrypt_to_local');
+            $container->removeDefinition('console.command.secrets_encrypt_from_local');
 
             return;
         }
 
         $loader->load('secrets.xml');
 
-        $container->setAlias(SecretStorageInterface::class, new Alias('secrets.storage.cache', false));
+        $container->getDefinition('secrets.vault')->replaceArgument(0, $config['vault_directory']);
 
-        $container->getDefinition('secrets.storage.files')->replaceArgument(0, $config['encrypted_secrets_dir']);
-        $container->getDefinition('secrets.encoder.sodium')->replaceArgument(0, $config['encryption_key']);
+        if (!$config['local_dotenv_file']) {
+            $container->removeDefinition('secrets.local_vault');
+        }
 
-        $container->registerForAutoconfiguration(SecretStorageInterface::class)
-            ->addTag('secret_storage');
+        if ($config['decryption_env_var']) {
+            $container->getDefinition('secrets.decryption_key')->replaceArgument(1, $config['decryption_env_var']);
+        } else {
+            $container->removeDefinition('secrets.decryption_key');
+        }
     }
 
     private function registerSecurityCsrfConfiguration(array $config, ContainerBuilder $container, XmlFileLoader $loader)