+++ src/Symfony/Bundle/FrameworkBundle/Secrets/SecretEnvVarProcessor.php
@@ -9,23 +9,26 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\FrameworkBundle\Secret;
+namespace Symfony\Bundle\FrameworkBundle\Secrets;
 
-use Symfony\Bundle\FrameworkBundle\Exception\SecretNotFoundException;
-use Symfony\Bundle\FrameworkBundle\Secret\Storage\SecretStorageInterface;
 use Symfony\Component\DependencyInjection\EnvVarProcessorInterface;
 use Symfony\Component\DependencyInjection\Exception\EnvNotFoundException;
 
 /**
  * @author Tobias Schultze <http://tobion.de>
+ * @author Nicolas Grekas <p@tchwork.com>
+ *
+ * @internal
  */
 class SecretEnvVarProcessor implements EnvVarProcessorInterface
 {
-    private $secretStorage;
+    private $vault;
+    private $localVault;
 
-    public function __construct(SecretStorageInterface $secretStorage)
+    public function __construct(AbstractVault $vault, AbstractVault $localVault = null)
     {
-        $this->secretStorage = $secretStorage;
+        $this->vault = $vault;
+        $this->localVault = $localVault;
     }
 
     /**
@@ -43,10 +46,14 @@ class SecretEnvVarProcessor implements EnvVarProcessorInterface
      */
     public function getEnv($prefix, $name, \Closure $getEnv)
     {
-        try {
-            return $this->secretStorage->getSecret($name);
-        } catch (SecretNotFoundException $e) {
-            throw new EnvNotFoundException($e->getMessage(), 0, $e);
+        if (null !== $this->localVault && null !== $secret = $this->localVault->reveal($name)) {
+            return $secret;
         }
+
+        if (null !== $secret = $this->vault->reveal($name)) {
+            return $secret;
+        }
+
+        throw new EnvNotFoundException($this->vault->getLastMessage() ?? sprintf('Secret "%s" not found or decryption key is missing.', $name));
     }
 }